backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.21
psycopg[binary]==3.2.1
psycopg2-binary==2.9.9
alembic==1.13.2
python-dotenv==1.0.1
pydantic-settings==2.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
